Move and refactor getOS function
