+++ frontend/packages/kubevirt-plugin/src/k8s/requests/vm/create/common.ts
@@ -1,40 +1,20 @@
 import { TemplateKind } from '@console/internal/module/k8s';
 import { getName, getNamespace } from '@console/shared/src';
 import { VMSettingsField } from '../../../../components/create-vm-wizard/types';
-import { getFieldValue } from '../../../../components/create-vm-wizard/selectors/vm-settings';
 import {
   ANNOTATION_DESCRIPTION,
+  APP,
   LABEL_USED_TEMPLATE_NAME,
   LABEL_USED_TEMPLATE_NAMESPACE,
   TEMPLATE_FLAVOR_LABEL,
   TEMPLATE_OS_LABEL,
   TEMPLATE_OS_NAME_ANNOTATION,
-  TEMPLATE_WORKLOAD_LABEL,
   TEMPLATE_VM_DOMAIN_LABEL,
   TEMPLATE_VM_NAME_LABEL,
-  APP,
+  TEMPLATE_WORKLOAD_LABEL,
 } from '../../../../constants/vm';
 import { VMWrapper } from '../../../wrapper/vm/vm-wrapper';
-import { getTemplateOperatingSystems } from '../../../../selectors/vm-template/advanced';
 import { VMTemplateWrapper } from '../../../wrapper/vm/vm-template-wrapper';
-import { operatingSystemsNative } from '../../../../components/create-vm-wizard/native/consts';
-import { concatImmutableLists, immutableListToShallowJS } from '../../../../utils/immutable';
-
-export const getOS = ({ vmSettings, iUserTemplates, openshiftFlag, iCommonTemplates }) => {
-  const operatingSystems = openshiftFlag
-    ? getTemplateOperatingSystems(
-        immutableListToShallowJS<TemplateKind>(
-          concatImmutableLists(iUserTemplates, iCommonTemplates),
-        ),
-      )
-    : operatingSystemsNative;
-
-  const osID = getFieldValue(vmSettings, VMSettingsField.OPERATING_SYSTEM);
-  return {
-    osID,
-    osName: (operatingSystems.find(({ id }) => id === osID) || {}).name,
-  };
-};
 
 export const initializeCommonMetadata = (
   settings: {
